Track summary progress by message id, not filtered count
Filtering the summarization input (dropping tool results and tool-only assistant
turns) broke the rolling-summary bookkeeping: summary_offset was stored as the
filtered message count, while the "needs summary?" check compared it against the
unfiltered user/assistant count. Every filtered session therefore looked
permanently behind and was reconsidered on every sweep — it never settled.

Switch summary_offset to a message-id high-water mark. A session needs work only
when user/assistant messages exist with an id beyond that mark; folding resumes
from those, and the mark advances to the latest id afterward (even when the only
new messages were filtered-out tool output, so the session still settles). This
makes the trigger and the filtered fold consistent and keeps incremental
maintenance cheap for active sessions.

+++ internal/types/types.go
@@ -19,9 +19,10 @@ type Session struct {
 	LastByteOffset int64   `json:"-"`
 	ImportedAt     string  `json:"imported_at,omitempty"`
 
-	// Rolling summary (Phase 5).
+	// Rolling summary (Phase 5). SummaryOffset is the message-id high-water mark
+	// the summary reflects.
 	Summary       string `json:"summary,omitempty"`
-	SummaryOffset int    `json:"-"`
+	SummaryOffset int64  `json:"-"`
 
 	// Durable cross-worktree identity, populated via LEFT JOIN session_identity.
 	RepoID       string `json:"repo_id,omitempty"`
